requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-groq
langgraph
fpdf2
gtts
python-multipart
typing-extensions
python-dotenv